# archive/MicroServices/Authentication_MicroService/src/swagger.json
{
  "swagger": "2.0",
   "schemes": [
    "http"
  ],
  "info": {
    "title": "DigiFabric Authentication Rest API",
    "version": "1.0"
  },
...
  "basePath": "/",
  "paths": {
...
    "/generateOtp": {
      "post": {
        "description": "The API fetches the token from previous call as a header along with the Content-Type. For channelProvider as sendgrid this API requires request body parameters as toRecipient(ex- someemail@gmail.com) and fromMail(ex- someemail@gmail.com). The response json will have details of nextCall and token with statusCode as 303.",
        "summary": "OTP would be either generated/skipped based on conditional hooks configured on the UI and the header values passed from the request. OTP would be generated if a conditional hook is added and same is passed through headers or if a conditional hook is not added at all. OTP would be skipped if a conditional hook is added and not passed through headers.",
        "operationId": "generateOtp",
        "tags": [
          "Hooks"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "303": {
            "description": "See Other",
            "schema": {
              "description": "",
              "type": "object",
              "properties": {
                "nextCall": {
                  "type": "string",
                  "minLength": 1
                },
                "token": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "nextCall",
                "token"
              ]
            }
          },
          "400": {
            "description": "Bad Request"
        },
        "401":{
          "description": "Not authorised",
          "schema": {
              "description": "",
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "error"
                
              ]
            }
        }
        },
        "security": [
          
        ],
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "description": "The token which is retained from the prior call should be passed as header.",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "Content-Type",
            "description": "Content-Type is the type of body.",
            "required": true,
            "type": "string",
            "enum": [
              "application/json"
            ]
          },
          {
            "in": "body",
            "name": "body",
            "description": "body of api call",
            "required": true,
            "schema": {
              "description": "",
              "type": "object",
              "properties": {
                "toRecipient": {
                  "type": "string",
                  "minLength": 1,
                  "enum": [
                    "drishyapillai0304@gmail.com"
                  ]
                },
                "fromMail": {
                  "type": "string",
                  "minLength": 1,
                  "enum": [
                    "Drishya.Pillai@cognizant.com"
                  ]
                }
              },
              "required": [
                "toRecipient",
                "fromMail"
              ]
            }
          }
        ]
      }
    },
...
  }
}